+++ R/prepare_predictions.R
@@ -138,9 +138,12 @@ prepare_predictions.brmsframe <- function(x, draws, sdata, ...) {
     if (any(ulapply(out$dpars[thetas], is.list))) {
       # theta was predicted
       missing_id <- which(ulapply(out$dpars[thetas], is.null))
-      out$dpars[[paste0("theta", missing_id)]] <- structure(
-        data2draws(0, c(ndraws, nobs)), predicted = TRUE
-      )
+      if (length(missing_id)) {
+        # a reference category exists; fill in its (zero) linear predictor
+        out$dpars[[paste0("theta", missing_id)]] <- structure(
+          data2draws(0, c(ndraws, nobs)), predicted = TRUE
+        )
+      }
     } else {
       # theta was not predicted
       out$dpars$theta <- do_call(cbind, out$dpars[thetas])