+++ tests/testthat/tests.families.R
@@ -101,6 +101,10 @@ test_that("mixture returns expected results and errors", {
                "Expecting at least 2 mixture components")
   expect_error(mixture(poisson, binomial, order = "x"),
                "Argument 'order' is invalid")
+  expect_true(is.na(mixture(gaussian, gaussian, refcat = NA)$refcat))
+  expect_null(mixture(gaussian, gaussian)$refcat)
+  expect_error(mixture(gaussian, gaussian, refcat = 1),
+               "'refcat' can only be NULL or NA")
 })
 
 test_that("response interval is defined correctly", {
@@ -121,3 +125,4 @@ test_that("default priors are as expected", {
 test_that("correct STAN code is used", {
     expect_equal(xbeta()$include, "fun_xbeta.stan")
 })
+